mybot/requirements.txt
alembic==1.10.2
APScheduler==3.6.3
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dropbox==11.36.0
ephem==4.1.4
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.38.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
ply==3.11
py==1.11.0
pyparsing==3.0.9
pytest==7.0.1
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.4
stone==3.3.1
termcolor==2.2.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
pycparser==2.20
PySocks==1.7.1
emoji==0.5.4